+++ core/src/main/java/com/mrivanplays/annotationconfig/core/PropertyConfig.java
@@ -1,8 +1,5 @@
 package com.mrivanplays.annotationconfig.core;
 
-import com.mrivanplays.annotationconfig.core.annotations.type.AnnotationType;
-import com.mrivanplays.annotationconfig.core.internal.AnnotatedConfigResolver;
-import com.mrivanplays.annotationconfig.core.internal.AnnotationHolder;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
@@ -12,38 +9,64 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
-import java.util.Set;
 
-/** Represents configuration, utilising .conf/.properties configuration type. */
+/**
+ * Represents configuration, utilising .conf/.properties configuration type.
+ *
+ * @since 1.0
+ * @author MrIvanPlays
+ */
 public final class PropertyConfig {
 
   private static final ValueWriter VALUE_WRITER = new PropertyValueWriter();
+
+  private static ConfigResolver configResolver;
+
+  /**
+   * Returns the {@link ConfigResolver} instance of PropertyConfig
+   *
+   * @return config resolver
+   */
+  public static ConfigResolver getConfigResolver() {
+    if (configResolver == null) {
+      generateConfigResolver();
+    }
+    return configResolver;
+  }
+
+  private static void generateConfigResolver() {
+    configResolver =
+        ConfigResolver.newBuilder()
+            .withCommentPrefix("# ")
+            .withValueWriter(VALUE_WRITER)
+            .withValueReader(
+                file -> {
+                  Properties properties = new Properties();
+                  try (Reader reader = new FileReader(file)) {
+                    properties.load(reader);
+                  } catch (IOException e) {
+                    throw new RuntimeException(e);
+                  }
+                  Map<String, Object> ret = new HashMap<>();
+                  for (Object key : properties.keySet()) {
+                    ret.put(String.valueOf(key), properties.get(key));
+                  }
+                  return ret;
+                })
+            .build();
+  }
 
   /**
    * Loads the config object from the file. If the file does not exist, it creates one.
    *
    * @param annotatedConfig annotated config
    * @param file file
+   * @deprecated use {@link #getConfigResolver()}. it has a much better description of methods. the
+   *     equivalent of this method there is {@link ConfigResolver#loadOrDump(Object, File, boolean)}
    */
+  @Deprecated
   public static void load(Object annotatedConfig, File file) {
-    Map<AnnotationHolder, Set<AnnotationType>> map =
-        AnnotatedConfigResolver.resolveAnnotations(annotatedConfig, false);
-    if (!file.exists()) {
-      AnnotatedConfigResolver.dump(annotatedConfig, map, file, "# ", VALUE_WRITER, false);
-      return;
-    }
-    Properties properties = new Properties();
-    try (Reader reader = new FileReader(file)) {
-      properties.load(reader);
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    Map<String, Object> toMap = new HashMap<>();
-    for (Object key : properties.keySet()) {
-      toMap.put(String.valueOf(key), properties.get(key));
-    }
-    AnnotatedConfigResolver.setFields(
-        annotatedConfig, toMap, map, "# ", VALUE_WRITER, file, true, false, false, null);
+    getConfigResolver().loadOrDump(annotatedConfig, file, true);
   }
 
   private static final class PropertyValueWriter implements ValueWriter {